requirements.txt
pdfplumber
python-docx
pyfiglet
regex

